server: Move createPublic and createPrivate
These should be in key_conversion.go
